Commit before code rebase

+++ MicroWars/Tesla.h
@@ -12,18 +12,18 @@ namespace microwars
 		float tesla_pos_x, tesla_pos_y;
 		float tesla_radius;
 		int tesla_health;
-		float tesla_x_factor;
+		int tesla_x_factor;
 
 		public:
 		
 		vector < int* > deleted_units;
 		
-		Tesla(float x, float y, float radius, float x_factor);
+		Tesla(float x, float y, float radius, int x_factor);
 		float return_tesla_pos(char option);
 		int check_unit_vicinity(float x, float y);
 		void change_health();
 		int return_health();
-		float return_x_factor();
+		int return_x_factor();
 
 		~Tesla();
 	};
